Move Google-related handler stuff to own package
